requirements.txt
Django==2.2.24
mongoengine==0.23.1
graphene_mongo==0.2.13
graphene_django==2.15.0
